Start formalizing the changes
