+++ collegamento/__init__.py
@@ -6,11 +6,11 @@ from .files_variant import FileClient, FileServer  # noqa: F401, E402
 from .simple_client_server import (  # noqa: F401, E402
     COMMANDS_MAPPING,
     USER_FUNCTION,
+    Client,
     CollegamentoError,
     Request,
     RequestQueueType,
     Response,
     ResponseQueueType,
-    SimpleClient,
-    SimpleServer,
+    Server,
 )
