+++ collegamento/simple_client_server/__init__.py
@@ -1,5 +1,5 @@
-from .client import SimpleClient, SimpleServer  # noqa: F401, E402
-from .misc import (  # noqa: F401, E402
+from .client import Client, Server  # noqa: F401, E402
+from .utils import (  # noqa: F401, E402
     COMMANDS_MAPPING,
     USER_FUNCTION,
     CollegamentoError,
